chromeos-base/chromeos-installer: avoid unnecessary mtd flash

+++ chromeos-base/chromeos-installer/files/flash_spi.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 LOADER_DIR=/usr/share/rock5b
 
@@ -8,10 +8,13 @@ main() {
  local zero="${LOADER_DIR}/zero.img"
  local loader="${LOADER_DIR}/spi.img"
 
- echo "installing $loader to $target"
- dd if=$zero of=$target status=progress bs=1024 conv=fdatasync,notrunc
- dd if=$loader of=$target status=progress bs=1024 conv=fdatasync,notrunc
- echo "done"
+ cmp "$loader" "$target"
+ if [ "$?" -ne 0 ]; then
+     echo "installing $loader to $target"
+ #    dd if=$zero of=$target status=progress bs=1024 conv=fdatasync,notrunc
+     dd if=$loader of=$target status=progress bs=1024 conv=fdatasync,notrunc
+     echo "done"
+ fi
 }
 
 main "$@"
